Создал новый файл и перенес туда MyHomePage
